Move tests into the correct location
