+++ components/ledger/ledger-utxo-flow/src/test/kotlin/net/corda/ledger/utxo/flow/impl/transaction/serializer/amqp/UtxoFilteredTransactionSerializerTest.kt
@@ -3,7 +3,7 @@ package net.corda.ledger.utxo.flow.impl.transaction.serializer.amqp
 import net.corda.internal.serialization.amqp.helper.TestSerializationService
 import net.corda.ledger.common.data.transaction.filtered.serializer.amqp.FilteredTransactionSerializer
 import net.corda.ledger.lib.utxo.flow.impl.transaction.filtered.UtxoFilteredTransactionImpl
-import net.corda.ledger.utxo.flow.impl.transaction.filtered.UtxoFilteredTransactionTestBase
+import net.corda.ledger.utxo.test.UtxoFilteredTransactionTestBase
 import net.corda.ledger.utxo.test.UtxoLedgerTest
 import net.corda.utilities.serialization.deserialize
 import net.corda.v5.ledger.utxo.transaction.filtered.UtxoFilteredTransaction
